requirements.txt
d20
discord.py
python-dotenv
lxml
requests
pytest
requests_mock
playwright
redis
psycopg2-binary
sqlalchemy
dataclasses-json